feat(#678): move discussion detail views
