+++ apps/web/src/app/[lang]/dho/[id]/@aside/discussions/[documentSlug]/page.tsx
@@ -1,14 +1,13 @@
 'use client';
 
-import { SidePanel } from '../../_components/side-panel';
-import { getDhoPathAgreements } from '../../../agreements/constants';
 import { Locale } from '@hypha-platform/i18n';
 import { useParams } from 'next/navigation';
 import { useDocumentSlug } from '@web/hooks/use-document-slug';
 import { useDocumentBySlug } from '@web/hooks/use-document-by-slug';
 import { useDiscussionByDocumentSlug } from '@web/hooks/use-discussion-by-document-slug';
 
-import { DocumentDetails, Chat } from '@hypha-platform/epics';
+import { DocumentDetails, Chat, SidePanel } from '@hypha-platform/epics';
+import { getDhoPathAgreements } from '../../../@tab/agreements/constants';
 
 type PageProps = {
   params: Promise<{ slug: string; id: string; lang: string }>;
